requirements.txt
async-timeout==5.0.1
certifi==2026.4.22
charset-normalizer==3.4.7
idna==3.15
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
requests==2.34.0
six==1.17.0
spotipy==2.26.0
urllib3==2.7.0
